linter refactoring
